split user into user and user admin
