+++ web/templates/base.html
@@ -15,7 +15,7 @@
         <nav class="navbar navbar-inverse navbar-fixed-top">
             <div class="container-fluid">
                 <div class="navbar-header">
-                    <a class="navbar-brand" href="/builds">
+                    <a class="navbar-brand" href="/">
                         <!--<img alt="Brand" src="...">-->
                         LeeroyCI
                     </a>
